+++ s/areas/cli/types.ts
@@ -11,7 +11,6 @@ export type CliConfig<C extends CommandTree> = {
 	readme?: string
 	columns?: number
 	theme?: ArgvTheme
-	shorthandBooleans?: boolean
 	onExit?: (code: number) => void
 	onHelp?: (help: string) => void
 	onMistake?: (mistake: string) => void