+++ s/examples/cafe.ts
@@ -1,7 +1,7 @@
 #!/usr/bin/env node
 
 import {cli} from "../areas/cli/cli.js"
-import {arg, choice, command, param} from "../areas/analysis/helpers.js"
+import {arg, choice, command, number, param, string} from "../areas/analysis/helpers.js"
 
 const input = cli(process.argv, {
 	name: "cafe",
@@ -14,11 +14,11 @@ const input = cli(process.argv, {
 		pizza: command({
 			help: `made with the finest mozza, mama mia!`,
 			args: [
-				arg("size").required(String, choice(["small", "medium", "large"])),
-				arg("crust").default(String, {fallback: "thick", ...choice(["thick", "thin"])}),
+				arg("size").required(string, choice(["small", "medium", "large"])),
+				arg("crust").default(string, "thick", choice(["thick", "thin"])),
 			],
 			params: {
-				slices: param.required(Number, {help: `we serve it by the slice`}),
+				slices: param.required(number, {help: `we serve it by the slice`}),
 				pepperoni: param.flag("p", {help: `
 					we go heavy on the meaty pepperoni
 				`}),
@@ -28,11 +28,11 @@ const input = cli(process.argv, {
 		icecream: command({
 			help: `it's actually gelatto!`,
 			args: [
-				arg("vessel").default(String, {fallback: "cone", help: `cone, waffle, or bowl?`}),
+				arg("vessel").default(string, "waffle", choice(["cone", "waffle", "bowl"])),
 			],
 			params: {
-				scoops: param.required(Number, {help: `how many scoops do you want?`}),
-				flavor: param.default(String, {fallback: "vanilla", help: `pick your favorite`}),
+				scoops: param.required(number, {help: `how many scoops do you want?`}),
+				flavor: param.default(string, "vanilla", {help: `pick your favorite`}),
 				"choco-dipped": param.flag("d", {help: `dunked for a chocolate coating`}),
 			},
 		}),