+++ s/tooling/obmap.ts
@@ -1,0 +1,17 @@
+
+export const obmap = <Ob extends object, Value>(
+		o: Ob,
+		transform: ObTransform<Ob, Value>,
+	) => {
+
+	return Object.fromEntries(
+		Object.entries(o).map(
+			([key, value]: any) => [key, transform(value, key)]
+		)
+	) as {[P in keyof Ob]: Value}
+}
+
+type ObTransform<Ob extends object, Value> = (
+	(value: Ob[keyof Ob], key: keyof Ob) => Value
+)
+