+++ changelog.md
@@ -1,4 +1,20 @@
 
+### v0.3.0
+- (breaking) custom types!
+    - import the types `import {string, number, boolean} from "@benev/argv"`
+    - `String` becomes `string`
+    - `Number` becomes `number`
+    - `Boolean` becomes `boolean`
+- (breaking) default fallback is no longer in the options
+    ```ts
+    // old
+    param.default(string, {fallback: "hello"})
+
+    // new
+    param.default(string, "hello")
+    ```
+- new helpers, `asType`, `asTypes`, `list`
+
 ### v0.2.0
 - (breaking) massive nuclear rewrite
 
@@ -11,3 +27,4 @@
 
 ### v0.0.0
 - initial release
+